Added return Types

+++ Project/Utility/OsUtility.py
@@ -5,7 +5,7 @@ import shutil
 class OsUtility:
 
     @staticmethod
-    def get_file_list(folder, ending=".png"):
+    def get_file_list(folder, ending=".png") -> list:
         # get all the png files in temporary folder <=> all pdf pages
         files = []
         for r, _, f in os.walk(folder):
@@ -16,22 +16,21 @@ class OsUtility:
         return files
 
     @staticmethod
-    def clean_up_folder(folder):
+    def clean_up_folder(folder) -> None:
         # removes the directory and files that were used in compression process
         print("--cleaning up--")
-        shutil.rmtree(folder)
+        if os.path.isdir(folder):
+            shutil.rmtree(folder)
 
     @staticmethod
-    def create_folder_if_not_exist(file_path):
-        # TODO proper support for directories
-        if not os.path.isdir(os.path.dirname(file_path)):
-            # file_path is a file
-            os.mkdir(os.path.dirname(file_path))
-        elif file_path[-1] == "/":
-            # file_path is a directory because it endswith /
+    def create_folder_if_not_exist(file_path) -> None:
+        # checks if .pdf file else creates folder if there is no folder, yet
+        if not file_path.endswith(".pdf") and not os.path.isdir(file_path):
             os.mkdir(file_path)
+        elif not os.path.isdir(os.path.dirname(file_path)):
+            os.mkdir(os.path.dirname(file_path))
 
     @staticmethod
-    def get_filename(full_path_to_file):
+    def get_filename(full_path_to_file) -> str:
         # remove .pdf, path (only Filename)
         return full_path_to_file[:-4].split(os.path.sep)[-1]
